0.1.0 - Initial implementation
This is... quite a bit more complete than I sometimes get around to for
a first pass. Did my best to flesh out the readme and everything.

Are there bugs? I don't know, probably. There usually are. After 5-6
hours of messing with this (and a not-insignificant portion of that
spent waiting for data imports to test different approaches' performance
& resilience to interruptions) I'm burned out for now. 😅