Refactor bulk DB loading with background mode
Without anything to say otherwise, `NPODB.bulk_add_NPOs()` would tend to
consume 100% of a CPU core. On multicore systems it isn't too bad, but
on a lower end machine with only 1 or 2 threads it could seriously
affect the bot's responsiveness. (Heck, it even affected responsiveness
of my test bot running in WSL1 on a Ryzen Z1 Extreme. Probably for I/O
reasons, not CPU, but still - ideally avoided.)

The function now chunks the iterator it's passed, and sleeps for a
second between batches if asked to run in `background` mode.

Launch code in `plugin` now decides contextually whether to run in
`background` mode or not. First DB load will run in foreground; periodic
updates run in the background with load-reducing `time.sleep()`s.

There is now also a `force` mode for `_check_pub78_updates_impl()` that
bypasses the `last_update` check, useful for testing etc.